refactor: constant 위치 변경, connection api return값 수정, gateway 수정
